Add (web) segments
