+++ apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/(one-segment)/[first]/otazka/[questionNumber]/page.tsx
@@ -1,6 +1,6 @@
 import { notFound } from "next/navigation";
 
-import { QuestionPage } from "../../../../../../components/server";
+import { QuestionPage } from "../../../../../../../components/server";
 
 export default async function Page({ params }: { params: Promise<{ questionNumber: string }> }) {
   const questionNumberInt = Number.parseInt((await params).questionNumber);
