+++ apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/(one-segment)/[first]/vysledek/page.tsx
@@ -1,0 +1,5 @@
+import { ResultPage } from "../../../../../../components/server";
+
+export default function Page() {
+  return <ResultPage />;
+}
