+++ apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/(two-segments)/[first]/[second]/layout.tsx
@@ -1,0 +1,16 @@
+import { loadCalculatorData } from "../../../../../../calculator/lib";
+import { CalculatorStoreProvider } from "../../../../../../calculator/stores";
+export default async function Layout({ children, params }: { children: React.ReactNode; params: Promise<{ first: string; second: string }> }) {
+  const { first, second } = await params;
+
+  const calculatorData = await loadCalculatorData({ key: first, group: second });
+
+  return (
+    <CalculatorStoreProvider calculatorData={calculatorData}>
+      <section>
+        <p>First: `{first}`</p>
+        <main>{children}</main>
+      </section>
+    </CalculatorStoreProvider>
+  );
+}
