+++ apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/(two-segments)/[first]/[second]/navod/page.tsx
@@ -1,0 +1,8 @@
+import { redirect } from "next/navigation";
+
+export default async function Page({ params }: { params: Promise<{ first: string; step: string }> }) {
+  const { first, step } = await params;
+  if (!step) {
+    redirect(`/${first}/navod/1`);
+  }
+}
