+++ apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/(two-segments)/[first]/[second]/otazka/[questionNumber]/page.tsx
@@ -1,0 +1,11 @@
+import { notFound } from "next/navigation";
+
+import { QuestionPage } from "../../../../../../../../components/server";
+
+export default async function Page({ params }: { params: Promise<{ questionNumber: string }> }) {
+  const questionNumberInt = Number.parseInt((await params).questionNumber);
+  if (Number.isNaN(questionNumberInt)) {
+    notFound();
+  }
+  return <QuestionPage questionNumber={questionNumberInt} />;
+}
