+++ apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/(two-segments)/[first]/[second]/otazka/page.tsx
@@ -1,0 +1,8 @@
+import { redirect } from "next/navigation";
+
+export default async function Page({ params }: { params: Promise<{ first: string; questionNumber: string }> }) {
+  const { first, questionNumber } = await params;
+  if (!questionNumber) {
+    redirect(`/${first}/otazka/1`);
+  }
+}
