+++ apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/(two-segments)/[first]/[second]/page.tsx
@@ -1,0 +1,6 @@
+import { redirect } from "next/navigation";
+
+export default async function Page({ params }: { params: Promise<{ first: string }> }) {
+  const { first } = await params;
+  redirect(`/${first}/navod`);
+}
