+++ apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/(two-segments)/[first]/[second]/rekapitulace/page.tsx
@@ -1,0 +1,5 @@
+import { ReviewPage } from "../../../../../../../components/server";
+
+export default function Page() {
+  return <ReviewPage />;
+}
