--- apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/[first]/[second]/navod/[step]/page.tsx
@@ -1,3 +1,0 @@
-import Page from "../../../navod/[step]/page";
-
-export default Page;
