--- apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/[first]/[second]/otazka/[questionNumber]/page.tsx
@@ -1,3 +1,0 @@
-import Page from "../../../otazka/[questionNumber]/page";
-
-export default Page;
