--- apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/[first]/[second]/rekapitulace/page.tsx
@@ -1,3 +1,0 @@
-import Page from "../../rekapitulace/page";
-
-export default Page;
