--- apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/[first]/[second]/vysledek/page.tsx
@@ -1,3 +1,0 @@
-import Page from "../../vysledek/page";
-
-export default Page;
