--- apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/[first]/layout.tsx
@@ -1,9 +1,0 @@
-export default async function Layout({ children, params }: { children: React.ReactNode; params: Promise<{ first: string }> }) {
-  const { first } = await params;
-  return (
-    <section>
-      <p>First: `{first}`</p>
-      <main>{children}</main>
-    </section>
-  );
-}
