--- apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/[first]/rekapitulace/page.tsx
@@ -1,5 +1,0 @@
-import { ReviewPage } from "../../../../../components/server";
-
-export default function Page() {
-  return <ReviewPage />;
-}
