--- apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/[first]/vysledek/page.tsx
@@ -1,5 +1,0 @@
-import { ResultPage } from "../../../../../components/server";
-
-export default function Page() {
-  return <ResultPage />;
-}
