+++ apps/www.volebnikalkulacka.cz/calculator/view-models/index.ts
@@ -1,1 +1,3 @@
+"use client";
+
 export * from "./calculator";
